feat(ui, session/worktree): Show `git diff` output in diff pane (#9)
This change adds the `git diff` of a worktree to the diff pane and also
updates the hardcoded values in the list UI to use real
addition/deletion numbers.

The diff is generated using `go-git` bindings and is updated
periodically (same as the refresh interval for the preview), and is
rendered in a viewport that supports scrolling in the case of long
diffs.

https://github.com/user-attachments/assets/3b72b181-6970-4549-a7fe-44750816456c