fix(app): use shift tab for auto yes (#40)
Claude code recently added the option to auto accept edits if you hit
shift+tab at the beginning of a session. This change updates auto yes
mode to press shift tab when a new session is made. It doesn't auto
accept bash command execution, so we still check the stdout for prompts
and tap enter as needed.

In https://github.com/smtg-ai/claude-squad/issues/39, we don't handle
the case where the "Yes, and don't ask again" option is gone. This
change updates the search text to search for "No, and tell claude what
to do differently".

Closes: https://github.com/smtg-ai/claude-squad/issues/39